+++ examples/Mixtures/slot_based.py
@@ -49,7 +49,7 @@ import pandas as pd
 from baybe.constraints import (
     DiscreteDependenciesConstraint,
     DiscretePermutationInvarianceConstraint,
-    DiscreteRepetitionConstraint,
+    DiscreteRepetitionLimitConstraint,
     DiscreteSumConstraint,
     ThresholdCondition,
 )
@@ -126,9 +126,9 @@ parameters = [
 # having two slots with the same substance or having only one slot with the combined
 # amounts. Thus, we want to make sure that there are no such duplicate label entries,
 # which can be achieved using a
-# {class}`~baybe.constraints.discrete.DiscreteRepetitionConstraint`:
+# {class}`~baybe.constraints.discrete.DiscreteRepetitionLimitConstraint`:
 
-no_duplicates_constraint = DiscreteRepetitionConstraint(
+no_duplicates_constraint = DiscreteRepetitionLimitConstraint(
     parameters=["Slot1_Label", "Slot2_Label", "Slot3_Label"],
     n_max_repetitions=1,
 )